+++ lib/src/widgets/library/library_widgets.dart
@@ -8,6 +8,7 @@ import 'package:path/path.dart' as p;
 import '../../core/layout_size.dart';
 import '../../core/playback_settings.dart';
 import '../../core/tag_rules.dart';
+import '../../models/library_sort.dart';
 import '../../models/platform_models.dart';
 import '../../models/video_item.dart';
 import '../../pages/player/player_dialog_content.dart';